Moved Util classes to correct package
