+++ src/main/scala/controller/prolog/util/PrologImplicits.scala
@@ -3,7 +3,7 @@ package controller.prolog.util
 import alice.tuprolog.{Struct, Term}
 import controller.prolog.structs.StructsNames.Predicates.StoryNodePredicateName
 import controller.prolog.structs.StructsNames.Predicates.Records.PathwayRecord
-import controller.util.serialization.StringUtil.RichString
+import controller.util.StringUtil.RichString
 import model.nodes.{Pathway, StoryNode}
 
 import scala.language.implicitConversions
